Bespoke Models

bespoke() defines a model that makes predictions using a set function. "Training" such a model only consists of verifying that the output is consistent with the types of outputs seen in the training data; the function is not tailored at all to better predict the training data. Such models can be useful for establishing a baseline which any true models should outperform.

Usage

bespoke(mode = "classification", engine = "bespoke", fn = NULL)

Arguments

mode

A single character string for the prediction outcome mode. Only "classification" is valid.

engine

A single character string specifying what computational engine to use for "fitting." Only "bespoke" is valid.

fn

A function that takes a data.frame as input and returns a vector (integer, character, or factor) indicating the outcomes as output (with one value per input row). We may someday extend this for probabilities.

Value

A specification for a model.