Prepare bespoke_classification Data for Prediction

Prepare bespoke_classification Data for Prediction

Usage

.predict_bespoke_classification_bridge(type, model, predictors)

Arguments

type

A single character. The type of predictions to generate. Valid options are:

"class" for "hard" class predictions.

"prob" for class probabilities.

model

A bespoke_classification model object. hardhat switches to model from here forward, I think because the object has to actually be a model at this point for anything to make sense, and they don't have to fight the predict generic for the name.

predictors

Forged predictor data.

Value

A tibble with output dependent on type.