Predict from a bespoke_classification

Predict from a bespoke_classification

Usage

# S3 method for class 'bespoke_classification'
predict(object, new_data, type = c("class", "prob"), ...)

Arguments

object

A bespoke_classification object.

new_data

A data frame or matrix of new predictors.

type

A single character. The type of predictions to generate. Valid options are:

"class" for "hard" class predictions.

"prob" for class probabilities.

...

Not used, but required for extensibility.

Value

A tibble of predictions. The number of rows in the tibble is guaranteed to be the same as the number of rows in new_data.